Do not try to symlink files that already exist.

+++ install.rb
@@ -1,10 +1,9 @@
 #!/usr/bin/ruby
 
-
 home = File.expand_path('~')
 
 Dir['*'].each do |file|
   next if file =~ /install/
   target = File.join(home, ".#{file}")
-  `ln -s #{File.expand_path file} #{target}`
+  `ln -s #{File.expand_path file} #{target}` unless File.exists?(target)
 end
